requirements.txt
# Bilibili Downloader GUI 依赖

# GUI框架
Pillow>=10.0.0

# B站API
bilibili-api-python>=16.1.0

# 异步HTTP
aiohttp>=3.8.0
aiofiles>=23.0.0

# 工具库
pydantic>=2.0.0
python-dotenv>=1.0.0
qrcode[pil]>=7.0.0
